# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Axix Finance API
  version: 0.2.0
  description: >-
    Public and admin endpoints for Axix Finance investment platform.
    This spec has been expanded with component schemas, standardized error shapes,
    and planned (future) admin observability endpoints annotated via `x-planned: true`.
servers:
  - url: https://axix-finance.fly.dev/api
    description: Production
  - url: http://localhost:8080/api
    description: Local Dev
...
security:
  - bearerAuth: []
paths:
...
  /auth/create-profile:
    post:
      tags: [Auth]
      summary: Create user profile (service role)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileRequest"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSummary"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateProfileRequest:
      type: object
      required: [email, username]
      properties:
        email: { type: string, format: email }
        username: { type: string, minLength: 3 }
    UserSummary:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        username: { type: string }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, nullable: true }
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              value:
                error: { code: UNAUTHORIZED, message: Authorization required }
...
    BadRequest:
      description: Invalid request payload/parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            badRequest:
              value:
                error: { code: BAD_REQUEST, message: Invalid input }
...